drm/radeon/kms: implement emulated_vblank_time in gpu driver
Prior to this commit the emulated_vblank_time field was implemented
in vcrtcm.  However, neither vcrtcm nor the pims make any use of
that field -- only the gpu driver uses it.  Further, implementing
the field in vcrtcm instead of in the gpu driver makes the code
(in particular, the locking scheme) more complex.  This commit
implements the field in the gpu driver.  A follow-on commit will
remove the implementation of the field from vcrtcm.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>