vcrtcm: release the spinlock before emulating vblank
We shouldn't be calling the vblank emulation in the GPU
with PCON spinlock held. That causes ugly deadlocks.
The purpose of the spinlock is just to make sure that
nobody pulled the rug underneath while we dereference
the PCON structure to extract the CRTC. Once we have
the CRTC it is safe to emulate the vblank even if
the PCON goes away in the mean time, as long as all
callbacks that GPU may make into the VCRTCM are
destruct-safe.

Since we are reworking the use of spinlocks, also
take logging function out of the critical section.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>