drm/radeon/kms: stick to blit copy in vcrtcm push
Starting with 3.8 kernel, the defauly copy function
in radeon driver uses asynchronous DMA. However, we need
to do some more testing with it before we decide to make
it the mechanism of choice for VCRTCM push.

This patch (temporarily) forces the use of known-working
blit copy until we are fully satisfied with the performance
and robustness of DMA copy.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>