drm/radeon/kms: invalidate pconid after calling vcrtcm_g_detach
In the current implementation of vcrtcm, the function vcrtcm_g_detach()
calls the gpu driver's detach callback.  This is a bug: Vcrtcm's gpu-side
api functions must not call the corresponding gpu callback.  (Similarly,
the pim-side functions must not call the pim callbacks.)  That bug will
be fixed in a follow-on commit.  To prepare for the bug fix, the radeon
driver must be modified to explicitly invalidate the pconid after calling
vcrtcm_g_detach().

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>