vcrtcm: do not ask pcon for fb status in push mode
In push mode, PCON can't really determine whether the
framebuffer copy is in progress or not, because all work
is essentially done by the GPU. GPU, on the other hand,
does not know the mode in which PCON operates until
it asks the VCRTCM module for status.

This patch makes vcrtcm_g_get_fb_status return early
(without bothering the PCON) if PCON is in push mode
and set the flag that essentially tells the GPU
that it is responsible for figurint out the status.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>