drm/radeon/kms: implement attach and post_attach callbacks
This commit implements the callbacks required by the new
vcrtcm-based attach ioctl.

NB: This is the first commit for which the vcrtcm-based ioctls
are fully supported.  It is also the last commit for which
the gpu-driver-based ioctls are supported.  Starting with the
next commit, only vcrtcm-based ioctls will be supported.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>