vcrtcm: add connectors to sysfs
This commit adds the pcon-connector relationships to sysfs.  Each attached
pcon has a symlink to its connector, and each connector has symlinks to
its attached pcons (currently at most one) and also to its drm sysfs entry.
Here is an example:

	|-- connectors
	|   `-- 0:69
	|       |-- pcons
	|       |   `-- 7 -> ../../pcons/7
	|       |-- drm_connector -> ../../../../../pci0000:00/0000:00:03.0/0000:02:00.0/0000:03:00.0/0000:04:00.0/drm/card0/card0-Virtual-16
	|-- pcons
	|   `-- 7
	|       |-- attach_minor
	|       |-- attached
	|       |-- connector -> ../../connectors/0:69
	.       .
	.       .
	.       .

Here pcon 7 is attached to connector 69 of the graphics card whose drm
minor is 0 (which typically corresponds to the device file /dev/dri/card0).
The sysfs file pcons/0/attach_minor contains the drm minor that was
specified to the attach operation; that minor is either the same as the
drm minor of the card, or the drm minor of a render node that was
constructed from that card.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>