v4l2pim: add system call counter
We have a subtle race condition between v4l2pim_release
function and all other system calls that touch videobuf queue.
To resolve it, we will need an indicator whether we are in a
(videobuf-affecting) system call or not. So each time we enter
a system call that touches videobuf we increment an atomic
counter and decrement it before we leave. The value
of the counter tells us how many videobuf-affecting
system calls are in progress at any time.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>