v4l2pim: add active status bit
New status bit indicates that there is an opener on the
video device. We use it for two purposes: First is
to ensure that there is only one opener at the time. Second
is to bail out of the system call early if there is a
closing (which may occur asynchronously) in progress.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>