vcrtcm: implement internal pconids
External PCON IDs can be rapidly reused as PCONs are created and destroyed.
If those IDs were used throughout VCRTCM, the GPU driver, and the PIMs,
then bugs could arise in which an operation is performed on the wrong PCON,
because both the intended (but now destroyed) PCON and the unintended PCON
have the same ID.  To prevent such aliasing bugs, each PCON is now also
assigned an internal ID, which is a 32-bit integer formed as the concatenation
of the 16-bit external ID and a 16-bit value that is incremented with each
new PCON.  Internal IDs are used throughout VCRTCM, the GPU driver, and the
PIMs.  When VCRTCM processes an IOCTL, VCRTCM first maps the external ID
within the IOCTL to the corresponding internal ID, and vice versa in the
return direction.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>