vcrtcm: check whether pimid would spill into OWNER bits
If an extremely large number of pims are registered (and possibly
unregistered), then the pimid would eventually spill over into
the VCRTCM_OWNER_XXX bits, which would lead to incorrect behavior
during allocations and deallocations.  The "right" way to prevent
this bug is to change the implementation to use a pim table, which
would allow for reuse of pimids.  However, because no one will ever
in practice register so many pims, we simply check whether the pimid
would spill into the VCRTCM_OWNER_XXX bits, in which case we refuse
to register the pim.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>