vcrtcm: add force option to destroy and detach ioctls
Prior to this commit, the pim was allowed to refuse to detach
a pcon.  This commit implements a 'force' option to the detach
ioctl, which forces the pim to detach the pcon.  Because the
destroy ioctl might first have to detach the pcon, the 'force'
option is also added to the destroy ioctl.  Note that this commit
breaks the build; the following sequence of commits will fix it.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>