vcrtcm: fix vblank-worker cancellation logic
Vcrtcm was calling cancel_delayed_work_sync() to cancel the vblank worker
while also holding the pconid lock.  This can lead to deadlocks; see the
code commentary for an explanation, and also the solution.  Note that the
solution requires changing the vcrtcm API --- specifically, the locking
requirements of vcrtcm_p_destroy().  Hence, this commit breaks the build.
The following sequence of commits will fix it.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>