v4l2pim: eliminate v4l2pim_thread
Instead of using one (quite busy) thread to deliver
frames, we do it in the v4l2pim_vblank context.

Besides reducing the CPU waste, the (beneficial) effect
of this patch is that it now delivers frames to a v4l2
device at the rate determined by the PCON FPS rate (which
is the correct behavior).

The application can just call read() in a loop and rest
assured that each frame will be delivered once and only once.
If the application is faster than the frame delivery, it will
block. If application is slower, frames will be dropped on
videobuf queue.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>