vcrtcm: rename device file and sysfs root
This commit:
	- renames /dev/pimmgr to /dev/vcrtcm
	- renames /sys/class/vcrtcm/pimmgr to /sys/class/vcrtcm/vcrtcm

Note the two "vcrtcm"s in /sys/class/vcrtcm/vcrtcm.  That is intentional.
The first "vcrtcm" is the class and the second is the device, corresponding
to /dev/vcrtcm.

Note also that this commit requires that corresponding changes be made
to all user-space tools and configuration files that access vcrtcm
via its device file or sysfs.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>