v4l2pim: factor out population of format structure
The code that populates the format structure (that
eventually makes it into the user space) is repeated
across three functions. Factor it out in one common
function.

The beneficial side-effect of this patch is that two
functions now also populate the pixelformat field which
they were forgetting to do before.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>