v4l2pim: add and use the function to retrieve fps
v4l2 device must know its FPS rate (so that v4l2 applications
can query it). This application comes from PCON structure, but
it must be queried while PCON is locked (because it's racing
with VCRTCM calls and callbacks).

This patch adds a helper function to query the FPS and makes
the use of it on the v4l2-facing side of the driver.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>