v4l2pim: shut a race window between instantiate and open
VCRTCM instantiate operation and V4L2 open system call are
racing: In theory, instantiate can create a minor (and thus
create a device that an application can open) before the
v4l2pim_minor associated v4l2pim_pcon structures are brought
in consistent state.

This patch shuts this race window by adding a new status
flag that indicates that minor is A-OK for use by the
V4L2-facing side of the driver. Until that flag is set, open()
system call fails.

We could have solved this by changing the order of certain
operations in v4l2pim_instantiate function, but that would
require major rework, so we take a pragmatic approach.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>