v4l2pim: snapshot framebuffer attributes at open time
Frame buffer size (on the VCRTCM side of the driver) can
change at any time (e.g. DRM application can change mode).
Videobuffer size, on the other hand, is queried by the
application when it sees fit (and used in the application
without the knowledge of the kernel). So the only safe
time to determine what buffer size we will use is
when the application opens the V4L2 device. From then, until
the the device is closed, we have to deal with the buffer
size that we determined at the time of open.

At this point, we still don't deal with the case when
DRM changes the framebuffer size on us in the middle of
a V4L2 session. This case will be dealt with in patches
to follow.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>