v4l2pim: handle different framebuffer and videobuf sizes
If DRM changes the frame buffer on us, while the V4L2 device
is open, we have to deal with this because we cannot change
the videobuf size on application if it does not want to close
and re-open the v4l2 device. So if the resize happens, we
have to either crop or pad: if dimension D of the framebuffer
becomes smaller, then we only fill the videobuf up to D (eaving
droppings in the unused section (that's padding). If, on the other
hand, dimension D of the framebuffer becomes larger, then we only
fill the videobuf up to the dimension D cutting of the portion
of the framebuffer (that's cropping).

This patch fixes a bug that made the video output look like
a broken TV when the frame buffer was resized.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>