vcrtcm: lock attached crtc before locking pconid
To prevent races between the drm ioctls and the gpu driver's vcrtcm-
registered callbacks, all places in which the pconid lock is taken
now first lock the attached crtc, if there is one.  The only exception
are the gpu-driver-side functions, which do *not* also lock the crtc,
because the drm ioctl logic already locks the crtc.  This change has
the effect of wrapping all of the gpu driver's vcrtcm-registered
callbacks with the per-crtc lock.  Thus, all races between the drm
ioctls and the gpu driver's vcrtcm-registered callbacks are now
eliminated.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>