drm: lock all CRTCs in syscalls that remove an fb
System calls that remove the frame buffer (by calling
drm_framebuffer_remove) can indirectly manipulate the
CRTC. The problem is that these system calls do not
have a-priori knowledge of which CRTC will be manipulated
(that will be determined by the underlying GPU driver).

Ugly as it seems, the way to eliminate the race is to
lock *all* CRTCs associated with the device before
removing the frame buffer.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>