vcrtcm: release ioctl mutex during pim deregistration
When crtc locking is implemented in a future commit, it will be
necessary that the vcrtcm_ioctl_mutex not be held during the pcon-
destruction phase of pim deregistration.  To prepare the ground for
that future commit, this commit releases that mutex during that phase.
To prevent new races between ioctls and pim deregistration, this
commit adds a being_deregistered field.  This commit, as an added
bonus, also improves ioctl error semantics: If the ioctl's pcon or
pim has disabled callbacks, the ioctl now fails.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>