vcrtcm: issue vblanks only if PCON DPMS state is on
Physical CRTCs stop progressing vblank-events if their
associated connector goes into DPMS-off state because
the clock source to the CRTC halts. Virtual CRTCs do not
behave this way, which is wrong. This patch fixes the
problem by stopping the associated vblank thread if
PCON is put in DPMS-off state. If we look at it through
"PCON is a connector" analogy, then this is the right
behavior.

The vblank-generation thread should be re-enabled when
we transition back from DPMS-off into DPMS-on state.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>